(Major) Creating the Transformer architecture #6: Created the "encoder_layer" and the "encoder" layers (taken from TF2.0 Transformer tutorial as-is)
SEE details of the encoder_layer here: https://www.tensorflow.org/alpha/tutorials/text/transformer#encoder_layer and the "encoder" here: https://www.tensorflow.org/alpha/tutorials/text/transformer#encoder .
The "encoder" is a wrapper-layer around the "encoder_layer" (I know, the naming is hard to follow .. :p).
Basically, the encoder_layer is the 'core' one that performs the multi-head attention and then the feed-forward network (grey-bordered area on the left in figure 1 in the original Vaswani et. al. paper).
The "Encoder" is a wrapper around the encoder_layer that firstly handles the input embedding and the positional encoding *and then* does the encoder_layers' stuff
(It's basically *everything* on the left in figure 1 in the original Vaswani et.al. "Att. is all you need" paper).