Created my own data pre-processing pipeline.
It is similar to the pipeline shown in the Portugese->English dataset here: https://www.tensorflow.org/alpha/tutorials/text/transformer#setup_input_pipeline
However, I decided to not use Tensorflow Dataset but instead rely on a Numpy-based approach and thus created my own pipeline.
Bear in mind feeding the model via a Tf.data flow is faster and I might eventually switch to that flow of feeding the model with the data.