Fixed the plot_attention feature in the transformer code.
Fix description: in the original TF2.0 Transformer tutorial's code - since it all was in one big jupyter nb file - the tokenizer objects were globally reachable for the plot func.
As I have moved the plotting func to a separate file I now needed to provide the reference to the tokenizers through the method arguments.